fix(build): keep randomly-seeded tests out of PIT's target suite
The first Deep-verification run aborted in PIT's coverage phase:
JsonModelBasedOracleTest.randomOperationSequenceMatchesOracle (seeded
with System.nanoTime) failed without mutation, and PIT refuses to run
against a non-green suite. Randomly-seeded exploration can legitimately
fail by discovering a real, previously unknown bug — inside PIT that
aborts the whole analysis without even printing the seed, so it belongs
in the regular test task where a failure reports the seed and full
operation log for exact reproduction.

* Extract the oracle into JsonModelOracleHarness; keep the fixed-seed
  (deterministic) runs in JsonModelBasedOracleTest and move the random
  repeats to a new JsonModelBasedOracleRandomTest.
* Pin pitest targetTests to deterministic classes only (fixed-seed
  oracle, diff regression test, concurrency harnesses); the jqwik
  property tests use random seeds per run and leave the PIT scope for
  the same reason.
* Document the green-suite constraint in docs/VERIFICATION.md.

The CI failure did not reproduce locally across ~140 fresh random
sequences; if it was a rare real seed rather than a PIT-minion resource
artifact, the random test will re-find it in normal CI with the seed
printed, to be added to the fixed-seed regression list.
